Move out pull request code from addContributor()
