+++ lib/add-contributor.js
@@ -1,10 +1,8 @@
 module.exports = addContributor;
 
+const { generatePrTitle } = require("./modules/helpers");
 const getUserDetails = require("./get-user-details");
-const ContentFiles = require("./modules/content-files");
-const convertMessage = require("commit-conv");
-
-const { generatePrTitle, generatePrBody }  = require("./modules/helpers");
+const handlePullRequest = require("./handle-pull-request");
 
 async function addContributor({
   context,
@@ -21,7 +19,7 @@ async function addContributor({
     username: who,
   });
 
-  if(config.isContributorWithRequestedContributionTypeAlreadyExists({login: who, contributions})) {
+  if (config.isContributorWithRequestedContributionTypeAlreadyExists({ login: who, contributions })) {
     // let user know in comment
     const message = `@${who} already contributed before to ${contributions.join(", ")}`;
 
@@ -38,55 +36,19 @@ async function addContributor({
     profile,
   });
 
-  // fetch all files that are configured in .all-contributors.rc ("files" key)
-  const contentFiles = new ContentFiles({
-    repository,
-  });
-  await contentFiles.fetch(config);
-  if (config.getOriginalSha() === undefined) {
-    contentFiles.init();
-  }
-  contentFiles.generate(config);
-  const filesByPathToUpdate = contentFiles.get();
-
-  // add the `.all-contributorsrc` config file to list of files to update in PR
-  filesByPathToUpdate[config.getPath()] = {
-    content: config.getRaw(),
-    originalSha: config.getOriginalSha(),
-  };
-
-  const convention = config.get().commitConvention;
-  const prTitle = convertMessage({
-    tag: "docs",
-    msg: generatePrTitle(`add ${who} as a contributor`, contributions),
-    convention
-  });
-
-  const skipCi = config.get().skipCi;
-  const prBodyMessage = `Adds @${who} as a contributor for ${contributions.join(
+  const prContentTitle = generatePrTitle(`add ${who} as a contributor`, contributions);
+  const prContentMessage = `Adds @${who} as a contributor for ${contributions.join(
     ", "
   )}.\n\nThis was requested by ${commentReply.replyingToWho()} [in this comment](${commentReply.replyingToWhere()})`
 
-  const prBody = generatePrBody(prBodyMessage, skipCi);
-
-  // create or update pull request
-  const {
-    pullRequestURL,
-    pullCreated,
-  } = await repository.createPullRequestFromFiles({
-    title: prTitle,
-    body: prBody,
-    filesByPath: filesByPathToUpdate,
+  const { pullRequestURL, pullCreated } = await handlePullRequest({
+    commentReply,
+    repository,
+    config,
+    prContentTitle,
+    prContentMessage,
     branchName,
-    convention,
   });
-
-  // let user know in comment
-  const message = pullCreated
-    ? `I've put up [a pull request](${pullRequestURL}) to add @${who}! :tada:`
-    : `I've updated [the pull request](${pullRequestURL}) to add @${who}! :tada:`;
-
-  commentReply.reply(message);
 
   return { pullRequestURL, pullCreated, user: { id, login } };
 }
